Renamed frontend file & made compiler class
