+++ main.py
@@ -9,7 +9,7 @@ A Big, Useful, & Efficient LaTeX Application
 
 # Import modules and classes
 from PyQt5.QtWidgets import QApplication
-from frontend import App
+from gui import App
 
 # Start the app
 app = QApplication([])
